Static cast size_t -> int64_t instead of vice versa for equals comparisons
    These were just introduced by a previous CL moving MemRef getRank to return int64_t. size_t could be smaller than 64 bits and in equals comparisons, signed vs unsigned doesn't matter. In these cases, we know right now that the particular int64_t is not larger than max size_t (because it currently comes directly from a size() call), the alternative cast plus equals comparison is always safe, so we might as well do it that way and no longer require reasoning deeper into the callstack.

    We are already assuming that size() calls fit into int64_t in a number of other cases like the aforementioned getRank() (since exabytes of RAM are rare). If we want to avoid this assumption we will have to come up with a principled way to do it throughout.

--

PiperOrigin-RevId: 250980297

+++ examples/Linalg/Linalg1/lib/ViewOp.cpp
@@ -40,7 +40,7 @@ void linalg::ViewOp::build(Builder *b, OperationState *result, Value *memRef,
                            ArrayRef<Value *> indexings) {
   MemRefType memRefType = memRef->getType().cast<MemRefType>();
   result->addOperands({memRef});
-  assert(indexings.size() == static_cast<size_t>(memRefType.getRank()) &&
+  assert(static_cast<int64_t>(indexings.size()) == memRefType.getRank() &&
          "unexpected number of indexings (must match the memref rank)");
 
   result->addOperands(indexings);
@@ -107,7 +107,7 @@ ParseResult linalg::ViewOp::parse(OpAsmParser *parser, OperationState *result) {
   if (!memRefType)
     return parser->emitError(parser->getNameLoc(),
                              "memRef type expected for first type");
-  if (indexingsInfo.size() != static_cast<size_t>(memRefType.getRank()))
+  if (static_cast<int64_t>(indexingsInfo.size()) != memRefType.getRank())
     return parser->emitError(parser->getNameLoc(),
                              "expected " + Twine(memRefType.getRank()) +
                                  " indexings");
@@ -116,7 +116,7 @@ ParseResult linalg::ViewOp::parse(OpAsmParser *parser, OperationState *result) {
     return parser->emitError(parser->getNameLoc(), "view type expected");
 
   ArrayRef<Type> indexingTypes = ArrayRef<Type>(types).drop_front().drop_back();
-  if (indexingTypes.size() != static_cast<size_t>(memRefType.getRank()))
+  if (static_cast<int64_t>(indexingTypes.size()) != memRefType.getRank())
     return parser->emitError(parser->getNameLoc(),
                              "expected " + Twine(memRefType.getRank()) +
                                  " indexing types");
